revert formatting changes revert buildSrc change, remove editorconfig
