--- .editorconfig
@@ -1,23 +1,0 @@
-[*]
-charset = utf-8
-end_of_line = lf
-tab_width = 2
-indent_size = 2
-indent_style = space
-insert_final_newline = false
-max_line_length = 120
-
-[*.java]
-max_line_length = 100
-ij_any_wrap_long_lines = true
-
-#todo discuss
-tab_width = 4
-indent_size = 4
-ij_continuation_indent_size = 8
-
-ij_java_imports_layout = $*,|,*
-ij_java_use_single_class_imports = true
-ij_java_layout_static_imports_separately = true
-ij_java_class_count_to_use_import_on_demand = 999
-ij_java_names_count_to_use_import_on_demand = 999
