+++ build-logic/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    `kotlin-dsl`
+}
+
+repositories {
+    gradlePluginPortal()
+}
+
+dependencies {
+    implementation(libs.plugin.indra)
+    implementation(libs.plugin.indra.git)
+}
