--- buildSrc/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-  `kotlin-dsl`
-}
-
-repositories {
-  gradlePluginPortal()
-}
-
-dependencies {
-  implementation(libs.plugin.indra)
-  implementation(libs.plugin.indra.git)
-}
