--- buildSrc/settings.gradle.kts
@@ -1,9 +1,0 @@
-dependencyResolutionManagement {
-  versionCatalogs {
-    create("libs") {
-      from(files("../gradle/libs.versions.toml"))
-    }
-  }
-}
-
-rootProject.name = "buildSrc"
